remove all support for language variants from API
it just wasn't working, and was making everything else overcomplex and unreadable.

Instead, variants of a language are treated as separate languages by the API.

We can still base our Peninsular Spanish rule-based transcriber off of the Pan-American one,
without copying all the rules.

We just replace the differing rules.


Former-commit-id: a3a37c7272494141fb919f600c65b58c64596dd0