German: improve vowel handling
write the somewhat complex rules (from Wikipedia) for when short and
long vowels occur.

These rules are not yet comprehensive, because there are many
unaccounted-for corner cases in the German spelling system --
such as with consonant sounds represented by 2+ letters (sch, ch),
or with syllable-final consonant clusters (-nd, -nt)

Need to find another resource to suggest whether long or short vowels
are more suitable as the default, when the spelling system is ambiguous.

These changes are as-yet untested, due to the unfinished work going on
to make this library also compile to JavaScript