fix(RuleBasedTranscriber): fix output after adding groups functionality
we originally tried to use provide the MatchingGroupCollection,
and regex capturing group functionality
(eg a \1 or $1 embedded in the output string)
by calling Regex.replace(String), or String.replace(Regex) --
they both seemed equivalent.

Unfortunately, this appeared to make the output crazy and repetetive.

Now, we instead pass the MatchingGroupCollection generated during
the initial Regex.find() call, makde to see if there's a match
at the beginning of the remaining input string.

The actual transferrence is back to being a simple full replacement
of the output string.

We can't use \1 / $1 anymore, but then we still haven't needed it yet,
and passing the MatchingGroupCollection as an argument to the lambda
is probably equivalent anyway, just more verbose (but Kotliny).